+++ Generator/Hints/HintTypes/ItemToItemPathHint.cs
@@ -32,7 +32,7 @@ namespace TPRandomizer.Hints
             Item srcItem,
             string tgtCheckName,
             bool srcUseDefiniteArticle = false,
-            Dictionary<int, byte> itemPlacements = null
+            Dictionary<int, int> itemPlacements = null
         )
         {
             this.srcItem = srcItem;
@@ -42,7 +42,7 @@ namespace TPRandomizer.Hints
             CalcDerived(genData, itemPlacements);
         }
 
-        private void CalcDerived(HintGenData genData, Dictionary<int, byte> itemPlacements)
+        private void CalcDerived(HintGenData genData, Dictionary<int, int> itemPlacements)
         {
             if (itemPlacements != null)
             {
@@ -119,7 +119,7 @@ namespace TPRandomizer.Hints
         public static ItemToItemPathHint decode(
             HintEncodingBitLengths bitLengths,
             BitsProcessor processor,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             Item srcItem = (Item)processor.NextByte();
