+++ Generator/Hints/HintTypes/JovaniRewardsHint.cs
@@ -46,7 +46,7 @@ namespace TPRandomizer.Hints
         public static JovaniRewardsHint decode(
             HintEncodingBitLengths bitLengths,
             BitsProcessor processor,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             int listSize = processor.NextInt(4);
@@ -170,7 +170,7 @@ namespace TPRandomizer.Hints
                 CheckStatusDisplay checkStatusDisplay,
                 bool useDefArticle = false,
                 bool isLogicalItem = false,
-                Dictionary<int, byte> itemPlacements = null
+                Dictionary<int, int> itemPlacements = null
             )
             {
                 this.checkName = checkName;
@@ -184,7 +184,7 @@ namespace TPRandomizer.Hints
                 CalcDerived(genData, itemPlacements);
             }
 
-            private void CalcDerived(HintGenData genData, Dictionary<int, byte> itemPlacements)
+            private void CalcDerived(HintGenData genData, Dictionary<int, int> itemPlacements)
             {
                 if (itemPlacements != null)
                 {
@@ -226,7 +226,7 @@ namespace TPRandomizer.Hints
             public static JovaniCheckInfo decode(
                 HintEncodingBitLengths bitLengths,
                 BitsProcessor processor,
-                Dictionary<int, byte> itemPlacements
+                Dictionary<int, int> itemPlacements
             )
             {
                 int checkId = processor.NextInt(bitLengths.checkId);
