+++ Generator/Hints/HintTypes/WothHint.cs
@@ -14,11 +14,7 @@ namespace TPRandomizer.Hints
         // always derived
         public Item item { get; private set; }
 
-        public WothHint(
-            AreaId areaId,
-            string checkName,
-            Dictionary<int, byte> itemPlacements = null
-        )
+        public WothHint(AreaId areaId, string checkName, Dictionary<int, int> itemPlacements = null)
         {
             this.areaId = areaId;
             this.checkName = checkName;
@@ -26,7 +22,7 @@ namespace TPRandomizer.Hints
             CalcDerived(itemPlacements);
         }
 
-        private void CalcDerived(Dictionary<int, byte> itemPlacements)
+        private void CalcDerived(Dictionary<int, int> itemPlacements)
         {
             if (itemPlacements != null)
             {
@@ -58,7 +54,7 @@ namespace TPRandomizer.Hints
         public static WothHint decode(
             HintEncodingBitLengths bitLengths,
             BitsProcessor processor,
-            Dictionary<int, byte> itemPlacements
+            Dictionary<int, int> itemPlacements
         )
         {
             AreaId areaId = AreaId.decode(bitLengths, processor);
