+++ Generator/Logic/CustomMsgData.cs
@@ -113,7 +113,7 @@ namespace TPRandomizer
 
         public static CustomMsgData Decode(
             SharedSettings sSettings,
-            Dictionary<int, byte> itemPlacements,
+            Dictionary<int, int> itemPlacements,
             string sixCharString
         )
         {
